# [Basic] `MaxMallocUsage` statistic should cover all zones