@xedin 2023-09-08
@swift-ci please smoke test compiler performance