@xedin 2023-09-08
Some grepping didn't find any uses of `malloc_create_zone` based in the sanitizer in compiler-rt lib so I think we are only using the default zone anyway.