@xedin 2023-09-11
# Summary for main smoketest

**Regressions found (see below)**

- [Debug](#debug-1acf7de8d2f7e3c7)
    - [debug brief](#debug-brief-1acf7de8d2f7e3c7)
    - [debug detailed](#debug-detailed-1acf7de8d2f7e3c7)
- [Release](#release-1acf7de8d2f7e3c7)
    - [release brief](#release-brief-1acf7de8d2f7e3c7)
    - [release detailed](#release-detailed-1acf7de8d2f7e3c7)



<a name="debug-1acf7de8d2f7e3c7">

# Debug

<a name="debug-brief-1acf7de8d2f7e3c7">

## debug brief
<details>
<summary>Regressed (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (2)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
Frontend.NumInstructionsExecuted | 1,643,758,170,321 | 1,647,557,285,165 | 3,799,114,844 | 0.23%
LLVM.NumLLVMBytesOutput | 40,195,760 | 40,196,528 | 768 | 0.0%
</details>

<a name="debug-detailed-1acf7de8d2f7e3c7">

## debug detailed
<details>
<summary>Regressed (10)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ModuleShadowCacheHit | 628 | 698 | 70 | 11.15% :no_entry:
AST.ModuleShadowCacheMiss | 149 | 152 | 3 | 2.01% :no_entry:
AST.NumSourceLinesPerSecond | 59,845 | 62,010 | 2,164 | 3.62% :no_entry:
Frontend.MaxMallocUsage | 0 | 14,025,441,368 | 14,025,441,368 | 100.0% :no_entry:
Sema.GetImplicitSendableRequest | 4,565 | 4,624 | 59 | 1.29% :no_entry:
Sema.NumRequirementMachineUnifiedConcreteTerms | 15,698 | 15,872 | 174 | 1.11% :no_entry:
Sema.ProtocolRequiresClassRequest | 734 | 747 | 13 | 1.77% :no_entry:
Sema.ResolveMacroRequest | 141 | 143 | 2 | 1.42% :no_entry:
Sema.StructuralRequirementsRequest | 791 | 801 | 10 | 1.26% :no_entry:
Sema.TypeAliasRequirementsRequest | 791 | 801 | 10 | 1.26% :no_entry:
</details>
<details>
<summary>Improved (0)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
</details>
<details>
<summary>Unchanged (delta < 1.0% or delta < 100.0ms) (319)</summary>

name | old | new | delta | delta_pct
---: | ---: | ---: | ---: | ---:
AST.ImportSetCacheHit | 61,256 | 61,272 | 16 | 0.03%
AST.ImportSetCacheMiss | 13,021 | 13,024 | 3 | 0.02%
AST.ImportSetFoldHit | 7,891 | 7,893 | 2 | 0.03%
AST.ImportSetFoldMiss | 5,130 | 5,131 | 1 | 0.02%
AST.ModuleVisibilityCacheHit | 334,034 | 334,034 | 0 | 0.0%
AST.ModuleVisibilityCacheMiss | 2,580 | 2,580 | 0 | 0.0%
AST.NumASTBytesAllocated | 907,652,388 | 909,634,735 | 1,982,347 | 0.22%
AST.NumASTScopeExpansions | 120,374 | 120,374 | 0 | 0.0%
AST.NumASTScopeLookups | 125,715 | 125,715 | 0 | 0.0%
AST.NumDecls | 6,712 | 6,712 | 0 | 0.0%
AST.NumDependencies | 4,048 | 4,048 | 0 | 0.0%
AST.NumIncrementalDependencies | 188 | 188 | 0 | 0.0%
AST.NumLinkLibraries | 0 | 0 | 0 | 0.0%
AST.NumLoadedModules | 3,971 | 3,971 | 0 | 0.0%
AST.NumLocalTypeDecls | 8 | 8 | 0 | 0.0%
AST.NumModuleLookupClassMember | 406 | 406 | 0 | 0.0%
AST.NumModuleLookupValue | 1,615,378 | 1,615,845 | 467 | 0.03%
AST.NumObjCMethods | 162 | 162 | 0 | 0.0%
AST.NumOperators | 6 | 6 | 0 | 0.0%
AST.NumPrecedenceGroups | 2 | 2 | 0 | 0.0%
AST.NumReferencedDynamicNames | 11 | 11 | 0 | 0.0%
AST.NumReferencedMemberNames | 324,323 | 324,323 | 0 | 0.0%
AST.NumReferencedTopLevelNames | 22,968 | 22,968 | 0 | 0.0%
AST.NumSourceBuffers | 7,564 | 7,564 | 0 | 0.0%
AST.NumSourceLines | 159,926 | 159,926 | 0 | 0.0%
AST.NumTotalClangImportedEntities | 35,403 | 35,451 | 48 | 0.14%
Frontend.NumInstructionsExecuted | 1,643,758,170,321 | 1,647,557,285,165 | 3,799,114,844 | 0.23%
Frontend.NumProcessFailures | 0 | 0 | 0 | 0.0%
IRGen.IRGenRequest | 532 | 532 | 0 | 0.0%
IRGen.OptimizedIRRequest | 0 | 0 | 0 | 0.0%
IRGen.SymbolObjectCodeRequest | 0 | 0 | 0 | 0.0%
IRModule.NumGOTEntries | 7,597 | 7,597 | 0 | 0.0%
IRModule.NumIRAliases | 3,141 | 3,141 | 0 | 0.0%
IRModule.NumIRBasicBlocks | 136,592 | 136,592 | 0 | 0.0%
IRModule.NumIRComdatSymbols | 0 | 0 | 0 | 0.0%
IRModule.NumIRFunctions | 78,166 | 78,166 | 0 | 0.0%
IRModule.NumIRGlobals | 91,498 | 91,498 | 0 | 0.0%
IRModule.NumIRIFuncs | 0 | 0 | 0 | 0.0%
IRModule.NumIRInsts | 1,257,132 | 1,257,132 | 0 | 0.0%
IRModule.NumIRNamedMetaData | 2,128 | 2,128 | 0 | 0.0%
IRModule.NumIRValueSymbols | 169,539 | 169,539 | 0 | 0.0%
LLVM.NumLLVMBytesOutput | 40,195,760 | 40,196,528 | 768 | 0.0%
Parse.ExportedSourceFileRequest | 0 | 0 | 0 | 0.0%
Parse.IDEInspectionSecondPassRequest | 0 | 0 | 0 | 0.0%
Parse.NumFunctionsParsed | 4,640 | 4,640 | 0 | 0.0%
Parse.NumIterableDeclContextParsed | 11,344 | 11,344 | 0 | 0.0%
Parse.ParseAbstractFunctionBodyRequest | 7,944 | 7,944 | 0 | 0.0%
Parse.ParseMembersRequest | 9,252 | 9,252 | 0 | 0.0%
Parse.ParseSourceFileRequest | 7,107 | 7,107 | 0 | 0.0%
Parse.ParseTopLevelDeclsRequest | 7,107 | 7,107 | 0 | 0.0%
SILGen.ASTLoweringRequest | 545 | 545 | 0 | 0.0%
SILGen.ParseSILModuleRequest | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILGenFunctions | 55,402 | 55,402 | 0 | 0.0%
SILModule.NumSILGenGlobalVariables | 4,099 | 4,099 | 0 | 0.0%
SILModule.NumSILGenVtables | 308 | 308 | 0 | 0.0%
SILModule.NumSILGenWitnessTables | 3,954 | 3,954 | 0 | 0.0%
SILModule.NumSILOptDefaultWitnessTables | 0 | 0 | 0 | 0.0%
SILModule.NumSILOptFunctions | 59,610 | 59,610 | 0 | 0.0%
SILModule.NumSILOptGlobalVariables | 4,136 | 4,136 | 0 | 0.0%
SILModule.NumSILOptVtables | 310 | 310 | 0 | 0.0%
SILModule.NumSILOptWitnessTables | 4,571 | 4,571 | 0 | 0.0%
SILOptimizer.ExecuteSILPipelineRequest | 2,699 | 2,699 | 0 | 0.0%
SILOptimizer.LoweredSILRequest | 0 | 0 | 0 | 0.0%
Sema.ABIMembersRequest | 4,258 | 4,258 | 0 | 0.0%
Sema.AbstractGenericSignatureRequest | 1,789 | 1,798 | 8 | 0.47%
Sema.AccessLevelRequest | 314,519 | 315,562 | 1,043 | 0.33%
Sema.ActorIsolationRequest | 107,138 | 107,181 | 43 | 0.04%
Sema.AllMembersRequest | 2,562 | 2,562 | 0 | 0.0%
Sema.AnyObjectLookupRequest | 17 | 17 | 0 | 0.0%
Sema.ApplyAccessNoteRequest | 59,078 | 59,080 | 2 | 0.0%
Sema.AreAllStoredPropertiesDefaultInitableRequest | 373 | 373 | 0 | 0.0%
Sema.AttachedPropertyWrapperTypeRequest | 567 | 567 | 0 | 0.0%
Sema.AttachedPropertyWrappersRequest | 367,080 | 367,903 | 823 | 0.22%
Sema.AttachedResultBuilderRequest | 26,995 | 26,995 | 0 | 0.0%
Sema.BodyInitKindRequest | 1,088 | 1,088 | 0 | 0.0%
Sema.BreakTargetRequest | 44 | 44 | 0 | 0.0%
Sema.CXXNamespaceMemberLookup | 0 | 0 | 0 | 0.0%
Sema.CallerSideDefaultArgExprRequest | 1,143 | 1,143 | 0 | 0.0%
Sema.CheckDistributedFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.CheckInconsistentAccessLevelOnImport | 1,064 | 1,064 | 0 | 0.0%
Sema.CheckInconsistentImplementationOnlyImportsRequest | 129 | 129 | 0 | 0.0%
Sema.CheckInconsistentSPIOnlyImportsRequest | 1,064 | 1,064 | 0 | 0.0%
Sema.CheckInconsistentWeakLinkedImportsRequest | 129 | 129 | 0 | 0.0%
Sema.CheckRedeclarationRequest | 30,889 | 30,889 | 0 | 0.0%
Sema.ClangCategoryLookupRequest | 0 | 0 | 0 | 0.0%
Sema.ClangDirectLookupRequest | 393 | 393 | 0 | 0.0%
Sema.ClangRecordMemberLookup | 393 | 393 | 0 | 0.0%
Sema.ClassAncestryFlagsRequest | 2,139 | 2,139 | 0 | 0.0%
Sema.ClosureEffectsRequest | 2,392 | 2,392 | 0 | 0.0%
Sema.ClosureHasExplicitResultRequest | 732 | 732 | 0 | 0.0%
Sema.CollectOverriddenDeclsRequest | 272,337 | 273,490 | 1,153 | 0.42%
Sema.CompareDeclSpecializationRequest | 16,080 | 16,129 | 49 | 0.3%
Sema.CompilerPluginLoadRequest | 0 | 0 | 0 | 0.0%
Sema.ConditionalRequirementsRequest | 14,356 | 14,363 | 7 | 0.05%
Sema.ConformanceHasEffectRequest | 0 | 0 | 0 | 0.0%
Sema.ConstantValueInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ContinueTargetRequest | 45 | 45 | 0 | 0.0%
Sema.CursorInfoRequest | 0 | 0 | 0 | 0.0%
Sema.CustomAttrNominalRequest | 99 | 99 | 0 | 0.0%
Sema.CustomAttrTypeRequest | 0 | 0 | 0 | 0.0%
Sema.CustomRefCountingOperation | 0 | 0 | 0 | 0.0%
Sema.CxxRecordAsSwiftType | 0 | 0 | 0 | 0.0%
Sema.CxxRecordSemantics | 1,163 | 1,163 | 0 | 0.0%
Sema.DefaultAndMaxAccessLevelRequest | 1,845 | 1,845 | 0 | 0.0%
Sema.DefaultArgumentExprRequest | 1,661 | 1,661 | 0 | 0.0%
Sema.DefaultArgumentInitContextRequest | 26 | 26 | 0 | 0.0%
Sema.DefaultArgumentTypeRequest | 475 | 475 | 0 | 0.0%
Sema.DefaultDefinitionTypeRequest | 128 | 128 | 0 | 0.0%
Sema.DefaultTypeRequest | 7,519 | 7,519 | 0 | 0.0%
Sema.DerivativeAttrOriginalDeclRequest | 0 | 0 | 0 | 0.0%
Sema.DifferentiableAttributeTypeCheckRequest | 0 | 0 | 0 | 0.0%
Sema.DirectLookupRequest | 1,015,785 | 1,016,271 | 486 | 0.05%
Sema.DirectOperatorLookupRequest | 41,354 | 41,362 | 8 | 0.02%
Sema.DirectPrecedenceGroupLookupRequest | 12,859 | 12,859 | 0 | 0.0%
Sema.DistributedModuleIsAvailableRequest | 0 | 0 | 0 | 0.0%
Sema.DynamicallyReplacedDeclRequest | 113,263 | 113,306 | 43 | 0.04%
Sema.EnumElementExprPatternRequest | 0 | 0 | 0 | 0.0%
Sema.EnumRawTypeRequest | 577 | 577 | 0 | 0.0%
Sema.EnumRawValuesRequest | 312 | 312 | 0 | 0.0%
Sema.ExistentialConformsToSelfRequest | 526 | 526 | 0 | 0.0%
Sema.ExpandAccessorMacros | 16,230 | 16,230 | 0 | 0.0%
Sema.ExpandChildTypeRefinementContextsRequest | 10,191 | 10,191 | 0 | 0.0%
Sema.ExpandExtensionMacros | 18,109 | 18,219 | 110 | 0.61%
Sema.ExpandMacroExpansionDeclRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMacroExpansionExprRequest | 0 | 0 | 0 | 0.0%
Sema.ExpandMemberAttributeMacros | 1,120,466 | 1,123,493 | 3,027 | 0.27%
Sema.ExpandPeerMacroRequest | 836,311 | 839,172 | 2,861 | 0.34%
Sema.ExpandSynthesizedMemberMacroRequest | 4,416 | 4,416 | 0 | 0.0%
Sema.ExprPatternMatchRequest | 1,042 | 1,042 | 0 | 0.0%
Sema.ExtendedNominalRequest | 10,476 | 10,476 | 0 | 0.0%
Sema.ExtendedTypeRequest | 3,146 | 3,146 | 0 | 0.0%
Sema.ExternalMacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.FragileFunctionKindRequest | 38,291 | 38,291 | 0 | 0.0%
Sema.FunctionOperatorRequest | 144 | 144 | 0 | 0.0%
Sema.GenericParamListRequest | 114,482 | 114,756 | 274 | 0.24%
Sema.GenericSignatureRequest | 52,279 | 52,311 | 31 | 0.06%
Sema.GetDestructorRequest | 359 | 359 | 0 | 0.0%
Sema.GetDistributedActorArgumentDecodingMethodRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorIDPropertyRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorImplicitCodableRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorInvocationDecoderRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorSystemPropertyRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedActorSystemRemoteCallFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedRemoteCallArgumentInitFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedRemoteCallTargetInitFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationDecoderDecodeNextArgumentFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordArgumentFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordErrorTypeFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationEncoderRecordReturnTypeFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedTargetInvocationResultHandlerOnReturnFunctionRequest | 0 | 0 | 0 | 0.0%
Sema.GetDistributedThunkRequest | 0 | 0 | 0 | 0.0%
Sema.GetSourceFileAsyncNode | 2 | 2 | 0 | 0.0%
Sema.GlobalActorAttributeRequest | 135,721 | 135,789 | 68 | 0.05%
Sema.GlobalActorInstanceRequest | 13 | 13 | 0 | 0.0%
Sema.HasCircularInheritedProtocolsRequest | 218 | 218 | 0 | 0.0%
Sema.HasCircularRawValueRequest | 308 | 308 | 0 | 0.0%
Sema.HasDefaultInitRequest | 1,526 | 1,526 | 0 | 0.0%
Sema.HasDynamicCallableAttributeRequest | 859 | 859 | 0 | 0.0%
Sema.HasDynamicMemberLookupAttributeRequest | 9,380 | 9,390 | 9 | 0.1%
Sema.HasImportsMatchingFlagRequest | 14,214 | 14,214 | 0 | 0.0%
Sema.HasInitAccessorRequest | 90,652 | 90,840 | 188 | 0.21%
Sema.HasIsolatedSelfRequest | 244,780 | 245,291 | 511 | 0.21%
Sema.HasMemberwiseInitRequest | 1,104 | 1,104 | 0 | 0.0%
Sema.HasMissingDesignatedInitializersRequest | 294 | 294 | 0 | 0.0%
Sema.HasSelfOrAssociatedTypeRequirementsRequest | 232 | 232 | 0 | 0.0%
Sema.HasStorageRequest | 33,464 | 33,464 | 0 | 0.0%
Sema.HasUserDefinedDesignatedInitRequest | 1,526 | 1,526 | 0 | 0.0%
Sema.IDEInspectionFileRequest | 0 | 0 | 0 | 0.0%
Sema.ImplementsAttrProtocolRequest | 0 | 0 | 0 | 0.0%
Sema.InferredGenericSignatureRequest | 2,399 | 2,399 | 0 | 0.0%
Sema.InheritedDeclsReferencedRequest | 131,541 | 131,867 | 326 | 0.25%
Sema.InheritedProtocolsRequest | 12,640 | 12,663 | 23 | 0.18%
Sema.InheritedTypeRequest | 5,509 | 5,519 | 10 | 0.18%
Sema.InheritsSuperclassInitializersRequest | 261 | 261 | 0 | 0.0%
Sema.InitAccessorPropertiesRequest | 959 | 959 | 0 | 0.0%
Sema.InitAccessorReferencedVariablesRequest | 0 | 0 | 0 | 0.0%
Sema.InitKindRequest | 2,507 | 2,507 | 0 | 0.0%
Sema.InterfaceTypeRequest | 395,978 | 396,999 | 1,021 | 0.26%
Sema.IsABICompatibleOverrideRequest | 4,608 | 4,608 | 0 | 0.0%
Sema.IsAccessorTransparentRequest | 8,009 | 8,009 | 0 | 0.0%
Sema.IsActorRequest | 41,489 | 41,582 | 93 | 0.23%
Sema.IsCallAsFunctionNominalRequest | 156 | 156 | 0 | 0.0%
Sema.IsDeclApplicableRequest | 0 | 0 | 0 | 0.0%
Sema.IsDeclRefinementOfRequest | 1,098 | 1,098 | 0 | 0.0%
Sema.IsDefaultActorRequest | 579 | 579 | 0 | 0.0%
Sema.IsDistributedActorRequest | 45,865 | 45,990 | 125 | 0.27%
Sema.IsDynamicRequest | 45,461 | 45,461 | 0 | 0.0%
Sema.IsFinalRequest | 76,186 | 76,427 | 241 | 0.32%
Sema.IsGetterMutatingRequest | 12,063 | 12,063 | 0 | 0.0%
Sema.IsImplicitlyUnwrappedOptionalRequest | 59,472 | 59,473 | 1 | 0.0%
Sema.IsMoveOnlyRequest | 11,647 | 11,647 | 0 | 0.0%
Sema.IsNonUserModuleRequest | 1,328 | 1,330 | 2 | 0.15%
Sema.IsObjCRequest | 41,150 | 41,152 | 2 | 0.0%
Sema.IsSafeUseOfCxxDecl | 0 | 0 | 0 | 0.0%
Sema.IsSetterMutatingRequest | 10,312 | 10,312 | 0 | 0.0%
Sema.IsSingleValueStmtRequest | 5,389 | 5,389 | 0 | 0.0%
Sema.IsStaticRequest | 21,567 | 21,567 | 0 | 0.0%
Sema.LazyStoragePropertyRequest | 16 | 16 | 0 | 0.0%
Sema.LocalDiscriminatorsRequest | 7,137 | 7,137 | 0 | 0.0%
Sema.LookupAllConformancesInContextRequest | 54,148 | 54,469 | 320 | 0.59%
Sema.LookupConformanceInModuleRequest | 1,376,006 | 1,376,872 | 865 | 0.06%
Sema.LookupInModuleRequest | 149,486 | 149,919 | 433 | 0.29%
Sema.LookupInfixOperatorRequest | 1,560 | 1,560 | 0 | 0.0%
Sema.LookupPostfixOperatorRequest | 0 | 0 | 0 | 0.0%
Sema.LookupPrecedenceGroupRequest | 520 | 520 | 0 | 0.0%
Sema.LookupPrefixOperatorRequest | 2 | 2 | 0 | 0.0%
Sema.MacroDefinitionRequest | 0 | 0 | 0 | 0.0%
Sema.MangleLocalTypeDeclRequest | 1 | 1 | 0 | 0.0%
Sema.MemberwiseInitPropertiesRequest | 959 | 959 | 0 | 0.0%
Sema.ModuleImplicitImportsRequest | 129 | 129 | 0 | 0.0%
Sema.ModuleLibraryLevelRequest | 0 | 0 | 0 | 0.0%
Sema.ModuleQualifiedLookupRequest | 55,651 | 56,082 | 431 | 0.77%
Sema.NamedLazyMemberLoadSuccessCount | 1,100,748 | 1,102,059 | 1,311 | 0.12%
Sema.NamingPatternRequest | 4,091 | 4,091 | 0 | 0.0%
Sema.NeedsNewVTableEntryRequest | 12,463 | 12,463 | 0 | 0.0%
Sema.NumAccessorBodiesSynthesized | 6,394 | 6,394 | 0 | 0.0%
Sema.NumAccessorsSynthesized | 7,412 | 7,412 | 0 | 0.0%
Sema.NumConformancePathsRecorded | 7,122 | 7,122 | 0 | 0.0%
Sema.NumConformancesDeserialized | 102,158 | 102,415 | 257 | 0.25%
Sema.NumConstraintScopes | 645,375 | 645,520 | 144 | 0.02%
Sema.NumConstraintsConsideredForEdgeContraction | 26,028 | 26,028 | 0 | 0.0%
Sema.NumCrossImportsChecked | 0 | 0 | 0 | 0.0%
Sema.NumCrossImportsFound | 0 | 0 | 0 | 0.0%
Sema.NumCyclicOneWayComponentsCollapsed | 0 | 0 | 0 | 0.0%
Sema.NumDeclsDeserialized | 1,536,956 | 1,541,631 | 4,675 | 0.3%
Sema.NumDeclsTypechecked | 45,229 | 45,229 | 0 | 0.0%
Sema.NumLazyIterableDeclContexts | 222,080 | 222,666 | 586 | 0.26%
Sema.NumLazyRequirementSignatures | 15,496 | 15,510 | 14 | 0.09%
Sema.NumLazyRequirementSignaturesLoaded | 11,340 | 11,351 | 10 | 0.09%
Sema.NumLeafScopes | 456,366 | 456,468 | 102 | 0.02%
Sema.NumRequirementMachineCompletionSteps | 401,919 | 402,630 | 711 | 0.18%
Sema.NumRequirementMachines | 34,998 | 35,088 | 89 | 0.26%
Sema.NumTypesDeserialized | 411,865 | 413,800 | 1,935 | 0.47%
Sema.NumUnloadedLazyIterableDeclContexts | 163,113 | 163,366 | 253 | 0.16%
Sema.ObjCInterfaceAndImplementationRequest | 21,265 | 21,265 | 0 | 0.0%
Sema.OpaqueReadOwnershipRequest | 7,906 | 7,906 | 0 | 0.0%
Sema.OpaqueResultTypeRequest | 0 | 0 | 0 | 0.0%
Sema.OperatorPrecedenceGroupRequest | 9 | 9 | 0 | 0.0%
Sema.OverriddenDeclsRequest | 103,399 | 103,712 | 313 | 0.3%
Sema.ParamSpecifierRequest | 31,330 | 31,330 | 0 | 0.0%
Sema.PatternBindingEntryRequest | 12,022 | 12,022 | 0 | 0.0%
Sema.PatternTypeRequest | 14,800 | 14,800 | 0 | 0.0%
Sema.PolymorphicEffectKindRequest | 15,224 | 15,224 | 0 | 0.0%
Sema.PolymorphicEffectRequirementsRequest | 0 | 0 | 0 | 0.0%
Sema.PotentialMacroExpansionsInContextRequest | 110,917 | 111,006 | 89 | 0.08%
Sema.PreCheckFunctionBodyRequest | 5,169 | 5,169 | 0 | 0.0%
Sema.PreCheckResultBuilderRequest | 0 | 0 | 0 | 0.0%
Sema.PreCheckReturnStmtRequest | 9,291 | 9,291 | 0 | 0.0%
Sema.PrimaryAssociatedTypesRequest | 288 | 289 | 1 | 0.35%
Sema.PrimarySourceFilesRequest | 129 | 129 | 0 | 0.0%
Sema.PropertyWrapperAuxiliaryVariablesRequest | 34,778 | 34,778 | 0 | 0.0%
Sema.PropertyWrapperBackingPropertyTypeRequest | 567 | 567 | 0 | 0.0%
Sema.PropertyWrapperInitializerInfoRequest | 10,095 | 10,095 | 0 | 0.0%
Sema.PropertyWrapperLValuenessRequest | 0 | 0 | 0 | 0.0%
Sema.PropertyWrapperMutabilityRequest | 11,967 | 11,967 | 0 | 0.0%
Sema.PropertyWrapperTypeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.ProtocolDependenciesRequest | 9,820 | 9,840 | 20 | 0.2%
Sema.ProtocolRequirementsRequest | 1,062 | 1,062 | 0 | 0.0%
Sema.ProvideDefaultImplForRequest | 272,337 | 273,490 | 1,153 | 0.42%
Sema.QualifiedLookupRequest | 141,967 | 142,081 | 114 | 0.08%
Sema.RangeInfoRequest | 0 | 0 | 0 | 0.0%
Sema.RawCommentRequest | 17,201 | 17,201 | 0 | 0.0%
Sema.RenamedDeclRequest | 0 | 0 | 0 | 0.0%
Sema.RequirementRequest | 3,286 | 3,286 | 0 | 0.0%
Sema.RequirementSignatureRequest | 12,131 | 12,152 | 20 | 0.17%
Sema.RequiresOpaqueAccessorsRequest | 30,084 | 30,084 | 0 | 0.0%
Sema.RequiresOpaqueModifyCoroutineRequest | 7,028 | 7,028 | 0 | 0.0%
Sema.ResolveEffectiveMemberwiseInitRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveImplicitMemberRequest | 14,714 | 14,714 | 0 | 0.0%
Sema.ResolveMacroConformances | 0 | 0 | 0 | 0.0%
Sema.ResolveProtocolNameRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveRawLayoutLikeTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeEraserTypeRequest | 0 | 0 | 0 | 0.0%
Sema.ResolveTypeRequest | 75,356 | 75,356 | 0 | 0.0%
Sema.ResultBuilderTypeRequest | 7,254 | 7,254 | 0 | 0.0%
Sema.ResultTypeRequest | 16,315 | 16,315 | 0 | 0.0%
Sema.RootAndResultTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.RootTypeOfKeypathDynamicMemberRequest | 0 | 0 | 0 | 0.0%
Sema.SPIGroupsRequest | 130,043 | 130,331 | 287 | 0.22%
Sema.ScopedImportLookupRequest | 0 | 0 | 0 | 0.0%
Sema.SelfAccessKindRequest | 13,865 | 13,865 | 0 | 0.0%
Sema.SelfBoundsFromGenericSignatureRequest | 2,784 | 2,784 | 0 | 0.0%
Sema.SelfBoundsFromWhereClauseRequest | 32,209 | 32,336 | 126 | 0.39%
Sema.SemanticAvailableRangeAttrRequest | 43 | 43 | 0 | 0.0%
Sema.SemanticBriefCommentRequest | 3,198 | 3,198 | 0 | 0.0%
Sema.SemanticUnavailableAttrRequest | 44,774 | 44,774 | 0 | 0.0%
Sema.SetterAccessLevelRequest | 3,381 | 3,381 | 0 | 0.0%
Sema.SimpleDidSetRequest | 64,668 | 64,793 | 124 | 0.19%
Sema.SpecializeAttrTargetDeclRequest | 78 | 78 | 0 | 0.0%
Sema.StorageImplInfoRequest | 30,135 | 30,135 | 0 | 0.0%
Sema.StoredPropertiesAndMissingMembersRequest | 1,122 | 1,122 | 0 | 0.0%
Sema.StoredPropertiesRequest | 6,421 | 6,421 | 0 | 0.0%
Sema.StructuralTypeRequest | 0 | 0 | 0 | 0.0%
Sema.SuperclassDeclRequest | 8,371 | 8,391 | 19 | 0.23%
Sema.SuperclassTypeRequest | 661 | 661 | 0 | 0.0%
Sema.SynthesizeAccessorRequest | 7,412 | 7,412 | 0 | 0.0%
…